MattParkerJigsawPuzzleProblem

Documentation for MattParkerJigsawPuzzleProblem.

In this YouTube video Matt Parker asks for a program to help design and solve jigsaw puzzles.

The idea is to produce a set of jigsaw puzzle pieces where the pieces can be assembled in more than one way, and, ideally, in exactly two ways.

This is an attempt at that.

Index

MattParkerJigsawPuzzleProblem.ALL_EDGE_TYPES

MattParkerJigsawPuzzleProblem.AbstractPuzzlePiece

MattParkerJigsawPuzzleProblem.Edge

MattParkerJigsawPuzzleProblem.EdgeType

MattParkerJigsawPuzzleProblem.GridCell

MattParkerJigsawPuzzleProblem.ImmutablePuzzlePiece

MattParkerJigsawPuzzleProblem.MutablePuzzlePiece

MattParkerJigsawPuzzleProblem.cardinal_directions_from

MattParkerJigsawPuzzleProblem.do_cardinal_directions

MattParkerJigsawPuzzleProblem.edge_direction

MattParkerJigsawPuzzleProblem.edge_index

MattParkerJigsawPuzzleProblem.edges_mate

MattParkerJigsawPuzzleProblem.new_grid

MattParkerJigsawPuzzleProblem.next

MattParkerJigsawPuzzleProblem.opposite

MattParkerJigsawPuzzleProblem.perimeter_edge_indices

MattParkerJigsawPuzzleProblem.perimeter_edge_indices

MattParkerJigsawPuzzleProblem.perimeters

MattParkerJigsawPuzzleProblem.piece_edge

MattParkerJigsawPuzzleProblem.previous

MattParkerJigsawPuzzleProblem.rotation

Descriptions

MattParkerJigsawPuzzleProblem.ALL_EDGE_TYPES — Constant

ALL_EDGE_TYPES

ALL_EDGE_TYPES is a vector of all of the EdgeTypes that have been created.

MattParkerJigsawPuzzleProblem.AbstractPuzzlePiece — Type

AbstractPuzzlePiece

AbstractPuzzlePiece is the abstract supertype for all types of jigsaw puzzle piece.

MattParkerJigsawPuzzleProblem.Edge — Type

Edge(::EdgeType, ::BallOrSocket)

Edge represents one edge of a puzzle piece. It has an edge_type. The bs field indicates whether the edge is a ball or socket.

MattParkerJigsawPuzzleProblem.EdgeType — Type

EdgeType(isperimeter::Bool)

Creates a unique EdgeType.

isperimeter indicates if the EdgeType is for an edge on the perimeter of the puzzle.

MattParkerJigsawPuzzleProblem.GridCell — Type

GridCell(::AbstractPuzzlePiece, rotation::Int)

A GridCell is the container for a puzzle piece in a puzzle grid.

MattParkerJigsawPuzzleProblem.ImmutablePuzzlePiece — Type

ImmutablePuzzlePiece(from::MutablePuzzlePiece)

Constructs an ImmutablePuzzlePiece from a MutablePuzzlePiece.

MattParkerJigsawPuzzleProblem.MutablePuzzlePiece — Type

MutablePuzzlePiece()

Constructs a MutablePuzzlePiece with no edges defined yet.

MattParkerJigsawPuzzleProblem.cardinal_directions_from — Method

cardinal_directions_from(d::CardinalDirection)

Returns the four caerdinal directions, starting with d, in next order.

MattParkerJigsawPuzzleProblem.do_cardinal_directions — Method

do_cardinal_directions(f; randomize=false)

Applies the function f to each of the four CardinalDirections.

If randomize is true then the directions are considered in random order.

MattParkerJigsawPuzzleProblem.edge_direction — Method

edge_direction(rotation::Int, index::Int)::CardinalDirection

For the specified edge_index and rotation of a puzzle piece, return the CardinalDirection that that edge faces.

MattParkerJigsawPuzzleProblem.edge_index — Method

edge_index(::Int)::Int

Turns the argument value into a valid edge index using the modulus operator. This allows us to use aritmetic on edge indices and still get a valid index.

MattParkerJigsawPuzzleProblem.edges_mate — Method

edges_mate(::Edge, ::Edge)::Bool

Two Edges mate if they have the same EdgeType and their bss are opposites.

MattParkerJigsawPuzzleProblem.new_grid — Method

new_grid(number_of_rows, number_of_columns)::Grid

Creates an empty puzzle grid of the specified dimensions.

MattParkerJigsawPuzzleProblem.next — Method

next(::CardinalDirection)::CardinalDirection

Returns the next direction clockwise from the given direction.

MattParkerJigsawPuzzleProblem.opposite — Method

opposite(::CardinalDirection)::CardinalDirection

Returns the opposite direction. N() and S() are opposites. E() and W() are opposites.

MattParkerJigsawPuzzleProblem.perimeter_edge_indices — Method

perimeter_edge_indices(p::AbstractPuzzlePiece)

Returns a vector of the indices of edges of the puzzle piece p that are perimeter edges.

MattParkerJigsawPuzzleProblem.perimeter_edge_indices — Method

perimeter_edge_indices(grid::Grid, row::int, col::Int)

Returns a vector of the indices of edges of the specified cell of grid that are perimeter edges.

MattParkerJigsawPuzzleProblem.perimeters — Method

perimeters(puzzle::MultipleSolutionPuzzle, row::Int, col::Int)

Returns a four element vector indicating for each direction (N, E, S, W) whether that side of the cell at row, col is a perimeter of the puzzle.

MattParkerJigsawPuzzleProblem.piece_edge — Method

piece_edge(puzzle_piece, index)::Edge

Returns the Edge of puzzle_piece which corresponds to the specified edge_index.

MattParkerJigsawPuzzleProblem.previous — Method

previous(::CardinalDirection)::CardinalDirection

Returns the previous direction, that which is counter-clockwise, from the given direction.

MattParkerJigsawPuzzleProblem.rotation — Method

rotation(rot::Int)

Normalizes the rotation of the placement of a puzzle piece to one of 0, 1, 2, or 3.