MattParkerJigsawPuzzleProblem

Documentation for MattParkerJigsawPuzzleProblem.

In this YouTube video Matt Parker asks for a program to help design and solve jigsaw puzzles.

The idea is to produce a set of jigsaw puzzle pieces where the pieces can be assembled in more than one way, and, ideally, in exactly two ways.

This is an attempt at that.

Index

MattParkerJigsawPuzzleProblem.ALL_EDGE_TYPES

MattParkerJigsawPuzzleProblem.Edge

MattParkerJigsawPuzzleProblem.EdgeType

MattParkerJigsawPuzzleProblem.SolvedPuzzle

MattParkerJigsawPuzzleProblem.SolvedPuzzlePiece

MattParkerJigsawPuzzleProblem.assign_perimeter_edges

MattParkerJigsawPuzzleProblem.assign_unique_unassigned_edges

MattParkerJigsawPuzzleProblem.do_cardinal_directions

MattParkerJigsawPuzzleProblem.edges_match

MattParkerJigsawPuzzleProblem.random_edge_types

Descriptions

MattParkerJigsawPuzzleProblem.ALL_EDGE_TYPES — Constant

ALL_EDGE_TYPES

ALL_EDGE_TYPES is a vector of all of the EdgeTypes that have been created.

MattParkerJigsawPuzzleProblem.Edge — Type

Edge(::EdgeType, ::BallOrSocket)

Edge represents one edge of a puzzle piece. It has an edge_type. The bs field indicates whether the edge is a ball or socket.

MattParkerJigsawPuzzleProblem.EdgeType — Type

EdgeType(isperimeter::Bool)

Creates a unique EdgeType.

isperimeter indicates if the EdgeType is for an edge on the perimeter of the puzzle.

MattParkerJigsawPuzzleProblem.SolvedPuzzle — Type

SolvedPuzzle(rows, columns)

Constructs a SolvedPuzzle with the specified numbers of rows and columns.

The grid field is populated with SolvedPuzzlePieces.

MattParkerJigsawPuzzleProblem.SolvedPuzzlePiece — Type

SolvedPuzzlePiece(row, column)

Represents the piece of a SolvedPuzzle at the specified row and column.

The edges field is a Dict which maps from a CardinalDirection to and Edge.

MattParkerJigsawPuzzleProblem.assign_perimeter_edges — Method

assign_perimeter_edges(SolvedPuzzle)::SolvedPuzzle

Assigns a unique EdgeType to the perimeter edge of each of the perimeter puzzle pieces.

The SolvedPuzzle is returned.

MattParkerJigsawPuzzleProblem.assign_unique_unassigned_edges — Method

assign_unique_unassigned_edges(sp::SolvedPuzzle)::SolvedPuzzle

Assigns an EdgeType] to each of the internal edges of the puzzle.

The SolvedPuzzle is returned.

MattParkerJigsawPuzzleProblem.do_cardinal_directions — Method

do_cardinal_directions(f; randomize=false)

Applies the function f to each of the four cardinal directions.

If randomize is true then the directions are considered in random order.

MattParkerJigsawPuzzleProblem.edges_match — Method

edges_match(::Edge, ::Edge)::Bool

Two Edges match if they have the same EdgeType and their bss are opposites.

MattParkerJigsawPuzzleProblem.random_edge_types — Method

random_edge_types(count, is_perimeter)

Create the specified number of unique EdgeTypes and returns them in a random order.